Solved a bug on sayial.asm
When the user of program try to enter 314.523 number for first number, then if he enter a number shorter than decimal part of the first number it will replace the empty sections from first number. So if he enter 7.9 for second number the second number will be 714.923. This bug occurs when the string is not cleaned(tamkatar, virgulkatar). So we add a section of code for clean the string every function call for enterNumber